+++ src/App.css
@@ -58,7 +58,7 @@
 img {
   width: 20.5rem;
   height: 7.5rem;
-  object-fit: contain;
+  /* object-fit: contain; */
 }
 
 .Restaurant {
