--- src/CartPage/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-// import { Container } from './styles';
-
-function CartPage() {
-  return <div>CartPage</div>;
-}
-
-export default CartPage;
