--- src/FeedPage/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-// import { Container } from './styles';
-
-function FeedPage() {
-  return <div>FeedPage</div>;
-}
-
-export default FeedPage;
