--- src/LoginPage/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-// import { Container } from './styles';
-
-function LoginPage() {
-  return <div>LoginPage</div>;
-}
-
-export default LoginPage;
