--- src/Pages/Cart.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import IndexCart from '../Components/Cart/index_Cart';
-
-function Cart() {
-    return(
-        <IndexCart />
-    );
-}
-
-export default Cart;
