--- src/Profile/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-// import { Container } from './styles';
-
-function Profile() {
-  return <div>Profile</div>;
-}
-
-export default Profile;
