--- src/SignUp/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-// import { Container } from './styles';
-
-function SignUp() {
-  return <div>SignUp</div>;
-}
-
-export default SignUp;
