+++ src/assets/images/logo-invert/logo-invert.png
